requirements.txt
google-cloud-storage==2.10.0
google-cloud-firestore==2.11.0
